requirements.txt
asyncio
aiofiles
aiohttp
flask
requests
python-telegram-bot
python-dotenv
pycryptodome
mysql-connector-python